Move trait impl check to impl encoder
