+++ prusti-encoder/src/encoders/ty/rust_ty.rs
@@ -658,8 +658,8 @@ pub enum Sizedness<'tcx> {
 impl<'tcx> Sizedness<'tcx> {
     fn map(self, f: impl FnOnce(ty::Ty<'tcx>) -> ty::Ty<'tcx>) -> Self {
         match self {
-            Sizedness::Dependent(param) => Sizedness::Dependent(f(param)),
-            sizedness => sizedness,
+            Sizedness::Dependent(ty) => Sizedness::Dependent(f(ty)),
+            other => other,
         }
     }
 }
